requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
numpy
PILLow
io
tensorflow
opencv-python